// src/app/store/action-types/group-conversation.types.ts
export enum GroupConversationActionType {
    GROUP_LOAD_CONVERSATION = '[Group Conversation] Load Group Conversation',
    GROUP_LOAD_CONVERSATION_SUCCESS = '[Group Conversation] Load Group Conversation Success',
    GROUP_LOAD_CONVERSATION_FAILURE = '[Group Conversation] Load Group Conversation Failure',

    GROUP_SEND_MESSAGE = '[Group Conversation] Send Conversation Message',
    GROUP_SEND_MESSAGE_SUCCESS = '[Group Conversation] Send Conversation Message Success',
    GROUP_SEND_MESSAGE_FAILURE = '[Group Conversation] Send Conversation Message Failure',

    GROUP_UPDATE_CONVERSATION = '[Group Conversation] Update Group Conversation',
    GROUP_UPDATE_CONVERSATION_SUCCESS = '[Group Conversation] Update Group Conversation Success',
    GROUP_UPDATE_CONVERSATION_FAILURE = '[Group Conversation] Update Group Conversation Failure',

    GROUP_CONVERSATION_LOADING_FALSE = '[Group Conversation] Conversation Loading False',
}